apps/ai-server/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# AI/ML Libraries
torch==2.5.1
transformers==4.46.3
diffusers==0.32.1
accelerate==1.1.1
pillow==11.0.0
sentencepiece==0.2.0

# vLLM for efficient LLM serving
vllm==0.6.6

# API utilities
python-multipart==0.0.17
httpx==0.28.1
aiofiles==24.1.0

# Data processing
numpy==2.1.3
